+++ src/basalstrength/MohrCoulombYieldStress.hh
@@ -37,7 +37,7 @@ public:
   virtual ~MohrCoulombYieldStress() = default;
 
   void set_till_friction_angle(const IceModelVec2S &input);
-private:
+protected:
   void restart_impl(const File &input_file, int record);
   void bootstrap_impl(const File &input_file, const YieldStressInputs &inputs);
   void init_impl(const YieldStressInputs &inputs);
@@ -51,6 +51,10 @@ private:
   void update_impl(const YieldStressInputs &inputs, double t, double dt);
 
   void finish_initialization(const YieldStressInputs &inputs);
+
+  IceModelVec2S m_till_phi;
+
+  std::shared_ptr<IceModelVec2T> m_delta;
 private:
   void till_friction_angle(const IceModelVec2S &bed_topography,
                            IceModelVec2S &result);
@@ -60,10 +64,6 @@ private:
                            const IceModelVec2S &ice_thickness,
                            const IceModelVec2CellType &cell_type,
                            IceModelVec2S &result);
-
-  IceModelVec2S m_till_phi;
-
-  std::shared_ptr<IceModelVec2T> m_delta;
 };
 
 } // end of namespace pism